-added liquids to background menu -added new functions to Rules editor -added rotators to MultiRotatorDrill icon -Gas and AdvancedContentInfo are now independent libraries -added rotatorsCount stat to MultiRotatorDrill

+++ tools/src/braindustry/tools/Generators.java
@@ -22,6 +22,7 @@ import arc.util.noise.Ridged;
 import arc.util.noise.VoronoiNoise;
 import braindustry.gen.ModContentRegions;
 import braindustry.type.ModUnitType;
+import braindustry.type.SelfIconGenerator;
 import mindustry.ctype.UnlockableContent;
 import mindustry.game.Team;
 import mindustry.gen.Legsc;
@@ -53,8 +54,6 @@ public class Generators {
 
         ObjectMap<Block, Pixmap> gens = new ObjectMap<>();
         Fi.get("../editor").mkdirs();
-        Fi.get("../override").mkdirs();
-        Fi.get("../override/editor").mkdirs();
         if (false) {
             generate("cracks", () -> {
                 for (int size = 1; size <= BlockRenderer.maxCrackSize; size++) {
@@ -118,7 +117,7 @@ public class Generators {
                     for (TextureRegion region : block.variantRegions()) {
                         GenRegion gen = (GenRegion) region;
                         if (gen.path == null) continue;
-                        gen.path.copyTo(Fi.get("../editor/editor-" + full( gen.path.name())));
+                        gen.path.copyTo(Fi.get("../editor/editor-" + gen.path.name()));
                     }
                 }
 
@@ -155,7 +154,7 @@ public class Generators {
 
                 try {
                     Pixmap last = null;
-                    if (block.outlineIcon && false) {
+                    if (block.outlineIcon) {
                         GenRegion region = (GenRegion) regions[block.outlinedIcon >= 0 ? block.outlinedIcon : regions.length - 1];
                         Pixmap base = get(region);
                         Pixmap out = last = base.outline(block.outlineColor, block.outlineRadius);
@@ -168,21 +167,26 @@ public class Generators {
                             }
                         }
 
-                        region.path.delete();
 
-                        save(out, block.name);
+                        if (false) {
+                            region.path.delete();
+                            save(out, block.name);
+                        }
                     }
 
                     if (!regions[0].found()) {
                         continue;
                     }
-
+                    boolean selfGenerator = block instanceof SelfIconGenerator;
                     Pixmap image = get(regions[0]);
 
                     int i = 0;
+
                     for (TextureRegion region : regions) {
                         i++;
-                        if (i != regions.length || last == null) {
+                        if (i == 1 && selfGenerator) {
+                            image.draw(((SelfIconGenerator) block).generate(get(regions[0]), ModImagePacker::get));
+                        } else if (i != regions.length || last == null) {
                             image.draw(get(region), true);
                         } else {
                             image.draw(last, true);
@@ -193,9 +197,8 @@ public class Generators {
                             image.draw(shardTeamTop, true);
                         }
                     }
-
-                    if (!(regions.length == 1 && regions[0] == Core.atlas.find(block.name) && shardTeamTop == null)) {
-                        save(image, "../override/block-" + full(block.name) + "-full");
+                    if (!(regions.length == 1 && regions[0] == Core.atlas.find(block.name) && shardTeamTop == null) || selfGenerator) {
+                        save(image, /*"block-" +*/ block.name + "-full");
                     }
 
                     save(image, "../editor/" + block.name + "-icon-editor");
@@ -203,7 +206,7 @@ public class Generators {
                     if (block.buildVisibility != BuildVisibility.hidden) {
                         saveScaled(image, block.name + "-icon-logic", logicIconSize);
                     }
-                    saveScaled(image, "../override/ui/block-" + full(block.name) + "-ui", Math.min(image.width, maxUiIcon));
+                    saveScaled(image, "../ui/block-" + block.name + "-ui", Math.min(image.width, maxUiIcon));
 
                     boolean hasEmpty = false;
                     Color average = new Color(), c = new Color();
@@ -374,7 +377,7 @@ public class Generators {
                             true);
                 }
 
-                save(image, "../override/unit-" + full(type.name) + "-full");
+                save(image, "unit-" + type.name + "-full");
 
                 Rand rand = new Rand();
                 rand.setSeed(type.name.hashCode());
@@ -415,7 +418,7 @@ public class Generators {
                 drawScaledFit(fit, image);
 
                 saveScaled(fit, type.name + "-icon-logic", logicIconSize);
-                save(fit, "../override/ui/unit-" + full(type.name) + "-ui");
+                save(fit, "../ui/unit-" + type.name + "-ui");
             } catch (Exception e) {
                 Log.err("WARNING: Skipping unit " + type.name + ": @", e);
             }
